Move pipreqs to uv-build-frontend
