--- crates/uv-analyze/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "uv-analyze"
-version = "0.0.1"
-edition = { workspace = true }
-
-[lib]
-doctest = false
-
-[lints]
-workspace = true
-
-[dependencies]
-uv-normalize = { workspace = true }
-
-ruff_python_ast = { workspace = true }
-ruff_python_parser = { workspace = true }
-rustc-hash = { workspace = true }
